Refactor to use new map loader. (dynamic map adapters)
