+++ src/components/map/index.tsx
@@ -1,5 +1,12 @@
-import MapboxMap from './components/map';
-import { Marker } from './components/marker';
+import { useAppConfig } from '@/hooks/use-app-config';
+import dynamic from 'next/dynamic';
 
-export default MapboxMap;
-export { Marker };
+export default function MapLoader(props) {
+  const appConfig = useAppConfig();
+  const mapAdapterName = appConfig.adapters.map;
+  const MapComponent = dynamic(() =>
+    import(`./${mapAdapterName}`).then((mod) => mod.default),
+  );
+
+  return <MapComponent {...props} />;
+}
