+++ src/components/map/mapbox/style.json
